feat: avoid duplicate computation for mxpx
