+++ arithmetization/src/main/java/net/consensys/linea/zktracer/module/mxp/moduleOperation/CancunMxpOperation.java
@@ -151,15 +151,15 @@ public class CancunMxpOperation extends LondonMxpOperation {
           .computation(true)
           .ct(i)
           .ctMax(cancunMxpCall.ctMax())
-          .pComputationWcpFlag(cancunMxpCall.exoCalls.get(i).wcpFlag())
-          .pComputationEucFlag(cancunMxpCall.exoCalls.get(i).eucFlag())
-          .pComputationExoInst(cancunMxpCall.exoCalls.get(i).instruction())
-          .pComputationArg1Hi(cancunMxpCall.exoCalls.get(i).arg1Hi())
-          .pComputationArg1Lo(cancunMxpCall.exoCalls.get(i).arg1Lo())
-          .pComputationArg2Hi(cancunMxpCall.exoCalls.get(i).arg2Hi())
-          .pComputationArg2Lo(cancunMxpCall.exoCalls.get(i).arg2Lo())
-          .pComputationResA(booleanToLong(cancunMxpCall.exoCalls.get(i).resultA()))
-          .pComputationResB(cancunMxpCall.exoCalls.get(i).resultB().toLong())
+          .pComputationWcpFlag(cancunMxpCall.exoCalls[i].wcpFlag())
+          .pComputationEucFlag(cancunMxpCall.exoCalls[i].eucFlag())
+          .pComputationExoInst(cancunMxpCall.exoCalls[i].instruction())
+          .pComputationArg1Hi(cancunMxpCall.exoCalls[i].arg1Hi())
+          .pComputationArg1Lo(cancunMxpCall.exoCalls[i].arg1Lo())
+          .pComputationArg2Hi(cancunMxpCall.exoCalls[i].arg2Hi())
+          .pComputationArg2Lo(cancunMxpCall.exoCalls[i].arg2Lo())
+          .pComputationResA(booleanToLong(cancunMxpCall.exoCalls[i].resultA()))
+          .pComputationResB(cancunMxpCall.exoCalls[i].resultB().toLong())
           .fillAndValidateRow();
     }
   }
